pyproject.toml
[tool.poetry]
name = "py-grpc-profile"
version = "0.1.0"
description = "profile for gRPC servers"
authors = ["Yoshiyuki HINO <yhinoz@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
importlib-metadata = {version = ">=0.12", python = "<3.8"}
grpcio = "^1.34.1"
pytest-grpc = "^0.8.0"
pytest-asyncio = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
pytest-cov = "^2.11.0"
black = "^20.8b1"
isort = {extras = ["pyproject"], version = "^5.7.0"}
flake8 = "^3.8.4"
mypy = "^0.790"
grpcio-tools = "^1.34.1"
snakeviz = "^2.1.0"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
src_paths = ["py_grpc_profile", "tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
